Refactor(user-module): apply updates to service methods
